git-svn-id: svn://70.113.44.214/VanaMirror/trunk@1034 09c1253f-3d09-8e40-9719-4320e56601b5

+++ src/ChannelServer/Mobs.h
@@ -81,7 +81,7 @@ namespace Mobs {
 
 class Mob {
 public:
-	Mob(int mapid, int mobid, Pos pos, int spawnid = -1, int fh = 0);
+	Mob(int mapid, int mobid, Pos pos, int spawnid = -1, short fh = 0);
 	void setPos(Pos pos) {
 		this->pos = pos;
 	}
